Updated windows templates to 1.3.3
V-Ray RT renamed to GPU and updated to Build 4.02.05. Blender-dr  needed a fix to work with the older image versions and the new versions due to a python now coming installed in the images.

+++ ncj/blender/scripts/setup-windows-pool.cmd
@@ -1,9 +1,14 @@
-
 echo # installing python ...
-choco install python -y 
-echo # refreshing environment vars ...
-call RefreshEnv.cmd
-echo # installing azure-batch sdk for python ...
+if exist "C:\Python37\python.exe" (
+    echo # python already installed
+    echo # installing microsoft build tools for the azure-batch cli
+    choco install microsoft-visual-cpp-build-tools -y
+) else (
+    choco install python -y 
+    echo # refreshing environment vars ...
+    call RefreshEnv.cmd
+    echo # installing azure-batch sdk for python ...
+)
 pip install azure-batch==4.1.3
 echo Exit Code is %errorlevel%
 exit /b %errorlevel%
